Added role functionality

+++ src/pages/Chapter.jsx
@@ -36,6 +36,7 @@ const Chapter = () => {
   const classService = new ClassService()
   const boardService = new BoardService()
   const subjectService = new SubjectService()
+  const role = localStorage.getItem("role")
 
   function initialLoad() {
     instuctorService.getAllInstocors().then((ins) => {
@@ -241,7 +242,7 @@ const Chapter = () => {
               Are you sure you want to delete {deleteObj?deleteObj.name:""}?<br></br>
               <Button onClick={() => {
                 setloading(true)
-                if (deleteObj) {
+                if (deleteObj && role=="admin") {
                   subjectService.deleteChapterID(deleteObj.subjectID, deleteObj.chapterID).then((res) => {
 
                     chapterService.deleteChapter(deleteObj.chapterID).then((res) => {
@@ -250,6 +251,9 @@ const Chapter = () => {
                       initialLoad()
                     })
                   })
+                }
+                else{
+                  alert("You are not authorized to perform this operation")
                 }
               }}>Yes</Button> <Button onClick={() => { setdeleteModal(false) }}>No</Button>
             </ModalBody>
